requirements.txt
streamlit==1.32.0
google-api-python-client==2.123.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-generativeai==0.4.0
pandas==2.1.4
matplotlib==3.8.2
pillow==10.2.0
python-dotenv==1.0.1
plotly==5.18.0
